Game: Fix various issues when server time is high
Bouncing items, fx_runner, bouncing missiles, movers, bouncing
objects, targets, triggers, blowing detpacks

+++ code/game/g_items.c
@@ -2113,7 +2113,7 @@ void G_BounceItem( gentity_t *ent, trace_t *trace ) {
 	int		hitTime;
 
 	// reflect the velocity on the trace plane
-	hitTime = level.previousTime + ( level.time - level.previousTime ) * trace->fraction;
+	hitTime = level.previousTime + (int)( ( level.time - level.previousTime ) * trace->fraction );
 	BG_EvaluateTrajectoryDelta( &ent->s.pos, hitTime, velocity );
 	dot = DotProduct( velocity, trace->plane.normal );
 	VectorMA( velocity, -2*dot, trace->plane.normal, ent->s.pos.trDelta );
